requirements.txt
-e .
aws-cdk.aws-iot
aws-cdk.aws-greengrass
boto3
botostubs